backend/requirements.txt
Flask
boto3
python-dotenv
Flask-Cors
psycopg2-binary